requirements.txt
streamlit
elasticsearch
python-dotenv
